# 3-cloud-deployer/src/api/credentials.py
"""
Permissions API Router

Provides endpoints for verifying cloud credentials against required permissions.
Supports AWS, Azure, and GCP credential validation.

Categories:
- "Permissions - Upload": Verify credentials from request body
- "Permissions - Project": Debug/local-cloud only. Verify credentials from project config files
"""

import os

from fastapi import APIRouter, HTTPException, Query
from pydantic import BaseModel, Field
from typing import Optional

from src.api.credentials_checker import (
    check_aws_credentials,
    check_aws_credentials_from_config,
)
from src.api.azure_credentials_checker import (
    check_azure_credentials,
    check_azure_credentials_from_config,
)
from src.api.gcp_credentials_checker import (
    check_gcp_credentials,
    check_gcp_credentials_from_config,
)
from src.api.error_models import ERROR_RESPONSES
from src.api.preflight import ProviderPreflightResponse, build_provider_preflight

router = APIRouter(prefix="/permissions")

LOCAL_CREDENTIAL_FILE_CHECKS_ENV = "ENABLE_LOCAL_CREDENTIAL_FILE_CHECKS"


def _require_local_credential_file_checks_enabled() -> None:
    if os.getenv(LOCAL_CREDENTIAL_FILE_CHECKS_ENV, "false").lower() != "true":
        raise HTTPException(
            status_code=403,
            detail={
                "error_code": "LOCAL_CREDENTIAL_FILE_CHECKS_DISABLED",
                "message": "File-based credential checks are disabled for this runtime.",
                "fix_suggestion": (
                    "Use POST /permissions/verify/{provider} with request-body credentials "
                    "or start the explicit local-cloud Compose override."
                ),
                "http_status": 403,
            },
        )


class AWSCredentialsRequest(BaseModel):
    """Request body for AWS credential validation."""

    aws_access_key_id: str = Field(..., description="AWS Access Key ID")
    aws_secret_access_key: str = Field(..., description="AWS Secret Access Key")
    aws_region: str = Field(..., description="AWS Region (e.g., 'eu-central-1')")
    aws_sso_region: Optional[str] = Field(
        None,
        description="AWS Region where IAM Identity Center (SSO) is enabled, if different from aws_region",
    )
    aws_session_token: Optional[str] = Field(
        None,
        description="Optional AWS Session Token for temporary credentials (from STS)",
    )


class AzureCredentialsRequest(BaseModel):
    """Request body for Azure credential validation."""

    azure_subscription_id: str = Field(..., description="Azure Subscription ID")
    azure_tenant_id: str = Field(..., description="Azure AD Tenant ID")
    azure_client_id: str = Field(
        ..., description="Service Principal Client/Application ID"
    )
    azure_client_secret: str = Field(..., description="Service Principal Client Secret")
    azure_region: str = Field(
        ..., description="Azure Region for general resources (e.g., 'italynorth')"
    )
    azure_region_iothub: str = Field(
        ...,
        description="Azure Region for IoT Hub (e.g., 'westeurope'), may differ from azure_region",
    )
    azure_region_digital_twin: str = Field(
        ...,
        description="Azure Region for Digital Twins (e.g., 'westeurope'), must be in ADT supported list",
    )


class GCPCredentialsRequest(BaseModel):
    """Request body for GCP credential validation."""

    # Dual-mode: either project_id (private account) OR billing_account (org account) required
    gcp_project_id: Optional[str] = Field(
        None, description="GCP Project ID for existing project (private accounts)"
    )
    gcp_billing_account: Optional[str] = Field(
        None,
        description="GCP Billing Account ID for auto-project creation (org accounts)",
    )
    gcp_credentials_file: str = Field(
        ..., description="Path to Service Account JSON key file"
    )
    gcp_region: str = Field(..., description="GCP Region (e.g., 'europe-west1')")


class CredentialsCheckResponse(BaseModel):
    """Response schema for credential validation."""

    status: str = Field(
        ...,
        description="Result status: 'valid', 'partial', 'invalid', 'check_failed', or 'error'",
    )
    message: str = Field(..., description="Human-readable result message")
    caller_identity: Optional[dict] = Field(
        None, description="AWS caller identity if credentials are valid"
    )
    can_list_policies: bool = Field(
        ..., description="Whether the credentials can list their own policies"
    )
    missing_check_permission: Optional[str] = Field(
        None,
        description="Permission missing to perform the check (if can_list_policies is False)",
    )
    by_layer: dict = Field(
        ...,
        description="Permission results organized by deployment layer (layer_1 through layer_5)",
    )
    by_service: dict = Field(
        ...,
        description="Permission results organized by AWS service with layer references",
    )
    summary: dict = Field(
        ..., description="Summary with total_required, valid, and missing counts"
    )


class AzureCredentialsCheckResponse(BaseModel):
    """Response schema for Azure credential validation."""

    status: str = Field(
        ...,
        description="Result status: 'valid', 'partial', 'invalid', 'check_failed', or 'error'",
    )
    message: str = Field(..., description="Human-readable result message")
    caller_identity: Optional[dict] = Field(
        None, description="Azure subscription/principal info"
    )
    can_list_roles: bool = Field(
        ..., description="Whether the credentials can list role assignments"
    )
    by_layer: dict = Field(
        ..., description="Permission results organized by deployment layer"
    )
    summary: dict = Field(
        ...,
        description="Summary with total_layers, valid_layers, partial_layers, invalid_layers counts",
    )
    recommended_roles: Optional[dict] = Field(
        None,
        description="Recommended roles: custom (preferred) and builtin alternatives",
    )


class GCPCredentialsCheckResponse(BaseModel):
    """Response schema for GCP credential validation."""

    status: str = Field(
        ...,
        description="Result status: 'valid', 'partial', 'invalid', 'sdk_missing', or 'error'",
    )
    message: str = Field(..., description="Human-readable result message")
    caller_identity: Optional[dict] = Field(
        None, description="GCP service account info"
    )
    project_access: Optional[dict] = Field(None, description="Project access status")
    api_status: Optional[dict] = Field(
        None, description="API enablement status by layer"
    )
    required_roles: list = Field(..., description="List of required IAM roles")


# ==========================================
# AWS Permissions - Upload
# ==========================================
@router.post(
    "/verify/aws",
    operation_id="verifyAwsCredentialsFromBody",
    response_model=CredentialsCheckResponse,
    tags=["Permissions - Upload"],
    summary="Verify AWS permissions from request body",
    description=(
        "**Purpose:** Verifies AWS credentials against all required permissions for the deployer.\n\n"
        "**When to call:** During wizard to validate user-provided AWS credentials.\n\n"
        "**Response:** Categorized results by layer (1-5) and by service."
    ),
    responses={
        200: {"description": "Credential validation completed (check status field)"},
        422: ERROR_RESPONSES[422],
        500: ERROR_RESPONSES[500],
    },
)
async def check_aws_from_body(request: AWSCredentialsRequest):
    """
    Validate AWS credentials from request body against all required permissions.

    Checks permissions for all 5 deployment layers:
    - **Layer 1**: IoT Core, Lambda (Dispatcher)
    - **Layer 2**: Lambda (Processor), Step Functions
    - **Layer 3**: DynamoDB, S3, EventBridge
    - **Layer 4**: IoT TwinMaker
    - **Layer 5**: Amazon Managed Grafana

    Returns a categorized list of valid and missing permissions.
    """
    return check_aws_credentials(request.model_dump())


@router.post(
    "/preflight/aws",
    operation_id="preflightAwsCredentialsFromBody",
    response_model=ProviderPreflightResponse,
    tags=["Permissions - Upload"],
    summary="Run AWS deployment preflight from request body",
    description=(
        "**Purpose:** Normalizes AWS credential validation into deployment-ready preflight checks.\n\n"
        "**When to call:** Before deployment to get actionable permission, account, and region failures.\n\n"
        "**Security:** Response is redacted and does not persist credentials."
    ),
    responses={
        200: {"description": "Preflight completed (check ready/status fields)"},
        422: ERROR_RESPONSES[422],
        500: ERROR_RESPONSES[500],
    },
)
async def preflight_aws_from_body(request: AWSCredentialsRequest):
    payload = request.model_dump()
    return build_provider_preflight("aws", check_aws_credentials(payload), payload)


# ==========================================
# AWS Permissions - Project
# ==========================================
@router.get(
    "/verify/aws",
    operation_id="verifyAwsCredentialsFromConfig",
    response_model=CredentialsCheckResponse,
    tags=["Permissions - Project"],
    summary="Verify AWS permissions from project config",
    description=(
        "**Purpose:** Debug/local-cloud only. Verifies AWS credentials from project's config_credentials.json.\n\n"
        "**Gate:** Disabled unless `ENABLE_LOCAL_CREDENTIAL_FILE_CHECKS=true` is set.\n\n"
        "**When to call:** To re-verify deployed project credentials.\n\n"
        "**Query param:** project (required) - project context to read."
    ),
    responses={
        200: {"description": "Credential validation completed (check status field)"},
        403: ERROR_RESPONSES[403],
        404: ERROR_RESPONSES[404],
        500: ERROR_RESPONSES[500],
    },
)
async def check_aws_from_config(
    project: Optional[str] = Query(
        None, description="Project name to load credentials from."
    ),
):
    """
    Validate AWS credentials from project's config_credentials.json.

    Reads the AWS credentials from the specified project's configuration file
    and validates them against all required permissions.

    A project is required because credential checks are request-scoped.
    """
    _require_local_credential_file_checks_enabled()
    return check_aws_credentials_from_config(project)


# ==========================================
# Azure Permissions - Upload
# ==========================================
@router.post(
    "/verify/azure",
    operation_id="verifyAzureCredentialsFromBody",
    response_model=AzureCredentialsCheckResponse,
    tags=["Permissions - Upload"],
    summary="Verify Azure permissions from request body",
    description=(
        "**Purpose:** Verifies Azure Service Principal credentials against required RBAC roles.\n\n"
        "**When to call:** During wizard to validate user-provided Azure credentials.\n\n"
        "**Checks:** Contributor and User Access Administrator role assignments."
    ),
    responses={
        200: {"description": "Credential validation completed (check status field)"},
        422: ERROR_RESPONSES[422],
        500: ERROR_RESPONSES[500],
    },
)
async def check_azure_from_body(request: AzureCredentialsRequest):
    """
    Validate Azure Service Principal credentials from request body.

    Checks RBAC role assignments for all deployment layers:
    - **Setup**: Resource Groups, Managed Identity, Storage
    - **Layer 0**: App Service Plan, Function Apps
    - **Layer 1**: IoT Hub, Event Grid, Role Assignments
    - **Layer 2**: Function Apps (Compute)
    - **Layer 3**: Cosmos DB, Blob Storage
    - **Layer 4**: Azure Digital Twins
    - **Layer 5**: Azure Managed Grafana

    Returns status and missing roles by layer.
    """
    return check_azure_credentials(request.model_dump())


@router.post(
    "/preflight/azure",
    operation_id="preflightAzureCredentialsFromBody",
    response_model=ProviderPreflightResponse,
    tags=["Permissions - Upload"],
    summary="Run Azure deployment preflight from request body",
    description=(
        "**Purpose:** Normalizes Azure credential validation into deployment-ready preflight checks.\n\n"
        "**When to call:** Before deployment to get actionable role, subscription, region, and secret-expiration failures.\n\n"
        "**Security:** Response is redacted and does not persist credentials."
    ),
    responses={
        200: {"description": "Preflight completed (check ready/status fields)"},
        422: ERROR_RESPONSES[422],
        500: ERROR_RESPONSES[500],
    },
)
async def preflight_azure_from_body(request: AzureCredentialsRequest):
    payload = request.model_dump()
    return build_provider_preflight("azure", check_azure_credentials(payload), payload)


# ==========================================
# Azure Permissions - Project
# ==========================================
@router.get(
    "/verify/azure",
    operation_id="verifyAzureCredentialsFromConfig",
    response_model=AzureCredentialsCheckResponse,
    tags=["Permissions - Project"],
    summary="Verify Azure permissions from project config",
    description=(
        "**Purpose:** Debug/local-cloud only. Verifies Azure credentials from project's config_credentials.json.\n\n"
        "**Gate:** Disabled unless `ENABLE_LOCAL_CREDENTIAL_FILE_CHECKS=true` is set.\n\n"
        "**When to call:** To re-verify deployed project credentials.\n\n"
        "**Query param:** project (required) - project context to read."
    ),
    responses={
        200: {"description": "Credential validation completed (check status field)"},
        403: ERROR_RESPONSES[403],
        404: ERROR_RESPONSES[404],
        500: ERROR_RESPONSES[500],
    },
)
async def check_azure_from_config(
    project: Optional[str] = Query(
        None, description="Project name to load credentials from."
    ),
):
    """
    Validate Azure credentials from project's config_credentials.json.

    Reads the Azure credentials from the specified project's configuration file
    and validates them against required RBAC role assignments.

    A project is required because credential checks are request-scoped.
    """
    _require_local_credential_file_checks_enabled()
    return check_azure_credentials_from_config(project)


# ==========================================
# GCP Permissions - Upload
# ==========================================
@router.post(
    "/verify/gcp",
    operation_id="verifyGcpCredentialsFromBody",
    response_model=GCPCredentialsCheckResponse,
    tags=["Permissions - Upload"],
    summary="Verify GCP permissions from request body",
    description=(
        "**Purpose:** Verifies GCP Service Account credentials against the selected versioned permission pack.\n\n"
        "**When to call:** During wizard to validate user-provided GCP credentials.\n\n"
        "**Checks:** Project access, billing, region, the fixed Phase 8 API baseline, project-testable IAM permissions, and explicit deferred resource/provider-operation checks."
    ),
    responses={
        200: {"description": "Credential validation completed (check status field)"},
        422: ERROR_RESPONSES[422],
        500: ERROR_RESPONSES[500],
    },
)
async def check_gcp_from_body(request: GCPCredentialsRequest):
    """
    Validate GCP Service Account credentials from request body.

    The active Phase 8 pack covers the standalone Six-layer v1 profile, including
    the PoC's custom GCP L4/L5 implementations. Permissions that cannot be
    evaluated safely on the project resource remain explicit warnings rather
    than false negatives; the fixed Phase 8 API superset is checked directly.
    """
    return check_gcp_credentials(request.model_dump())


@router.post(
    "/preflight/gcp",
    operation_id="preflightGcpCredentialsFromBody",
    response_model=ProviderPreflightResponse,
    tags=["Permissions - Upload"],
    summary="Run GCP deployment preflight from request body",
    description=(
        "**Purpose:** Normalizes GCP credential validation into deployment-ready preflight checks.\n\n"
        "**When to call:** Before deployment to get actionable project, billing, API, and region failures.\n\n"
        "**Security:** Response is redacted and does not persist credentials."
    ),
    responses={
        200: {"description": "Preflight completed (check ready/status fields)"},
        422: ERROR_RESPONSES[422],
        500: ERROR_RESPONSES[500],
    },
)
async def preflight_gcp_from_body(request: GCPCredentialsRequest):
    payload = request.model_dump()
    return build_provider_preflight("gcp", check_gcp_credentials(payload), payload)


# ==========================================
# GCP Permissions - Project
# ==========================================
@router.get(
    "/verify/gcp",
    operation_id="verifyGcpCredentialsFromConfig",
    response_model=GCPCredentialsCheckResponse,
    tags=["Permissions - Project"],
    summary="Verify GCP permissions from project config",
    description=(
        "**Purpose:** Debug/local-cloud only. Verifies GCP credentials from project's config_credentials.json.\n\n"
        "**Gate:** Disabled unless `ENABLE_LOCAL_CREDENTIAL_FILE_CHECKS=true` is set.\n\n"
        "**When to call:** To re-verify deployed project credentials.\n\n"
        "**Query param:** project (required) - project context to read."
    ),
    responses={
        200: {"description": "Credential validation completed (check status field)"},
        403: ERROR_RESPONSES[403],
        404: ERROR_RESPONSES[404],
        500: ERROR_RESPONSES[500],
    },
)
async def check_gcp_from_config(
    project: Optional[str] = Query(
        None, description="Project name to load credentials from."
    ),
):
    """
    Validate GCP credentials from project's config_credentials.json.

    Reads the GCP credentials from the specified project's configuration file
    and validates them against required permissions.

    A project is required because credential checks are request-scoped.
    """
    _require_local_credential_file_checks_enabled()
    return check_gcp_credentials_from_config(project)