# docs/contracts/openapi/deployer.openapi.json
{
  "components": {
    "schemas": {
      "AWSCredentialsRequest": {
        "description": "Request body for AWS credential validation.",
        "properties": {
          "aws_access_key_id": {
            "description": "AWS Access Key ID",
            "title": "Aws Access Key Id",
            "type": "string"
          },
          "aws_region": {
            "description": "AWS Region (e.g., 'eu-central-1')",
            "title": "Aws Region",
            "type": "string"
          },
          "aws_secret_access_key": {
            "description": "AWS Secret Access Key",
            "title": "Aws Secret Access Key",
            "type": "string"
          },
          "aws_session_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional AWS Session Token for temporary credentials (from STS)",
            "title": "Aws Session Token"
          },
          "aws_sso_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AWS Region where IAM Identity Center (SSO) is enabled, if different from aws_region",
            "title": "Aws Sso Region"
          }
        },
        "required": [
          "aws_access_key_id",
          "aws_secret_access_key",
          "aws_region"
        ],
        "title": "AWSCredentialsRequest",
        "type": "object"
      },
...
      "AzureCredentialsRequest": {
        "description": "Request body for Azure credential validation.",
        "properties": {
          "azure_client_id": {
            "description": "Service Principal Client/Application ID",
            "title": "Azure Client Id",
            "type": "string"
          },
          "azure_client_secret": {
            "description": "Service Principal Client Secret",
            "title": "Azure Client Secret",
            "type": "string"
          },
          "azure_region": {
            "description": "Azure Region for general resources (e.g., 'italynorth')",
            "title": "Azure Region",
            "type": "string"
          },
          "azure_region_digital_twin": {
            "description": "Azure Region for Digital Twins (e.g., 'westeurope'), must be in ADT supported list",
            "title": "Azure Region Digital Twin",
            "type": "string"
          },
          "azure_region_iothub": {
            "description": "Azure Region for IoT Hub (e.g., 'westeurope'), may differ from azure_region",
            "title": "Azure Region Iothub",
            "type": "string"
          },
          "azure_subscription_id": {
            "description": "Azure Subscription ID",
            "title": "Azure Subscription Id",
            "type": "string"
          },
          "azure_tenant_id": {
            "description": "Azure AD Tenant ID",
            "title": "Azure Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "azure_subscription_id",
          "azure_tenant_id",
          "azure_client_id",
          "azure_client_secret",
          "azure_region",
          "azure_region_iothub",
          "azure_region_digital_twin"
        ],
        "title": "AzureCredentialsRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HttpErrorEnvelope": {
        "description": "Actual FastAPI error shape returned by HTTPException handlers.",
        "example": {
          "detail": "The requested operation could not be completed."
        },
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "description": "Human-readable error text or a structured operation-specific error. Clients must preserve structured detail objects when present.",
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HttpErrorEnvelope",
        "type": "object"
      },
...
      "ProviderPreflightResponse": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/ProviderPreflightCheck"
            },
            "title": "Checks",
            "type": "array"
          },
          "provider": {
            "enum": [
              "aws",
              "azure",
              "gcp"
            ],
            "title": "Provider",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "passed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "ready",
          "status",
          "summary",
          "checks"
        ],
        "title": "ProviderPreflightResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Internal API for deploying, destroying, and inspecting Digital Twin resources.",
    "title": "Digital Twin Manager API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/permissions/preflight/aws": {
      "post": {
        "description": "**Purpose:** Normalizes AWS credential validation into deployment-ready preflight checks.\n\n**When to call:** Before deployment to get actionable permission, account, and region failures.\n\n**Security:** Response is redacted and does not persist credentials.",
        "operationId": "preflightAwsCredentialsFromBody",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AWSCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPreflightResponse"
                }
              }
            },
            "description": "Preflight completed (check ready/status fields)"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The request body failed validation."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Validation Error - The request body failed schema validation"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Run AWS deployment preflight from request body",
        "tags": [
          "Permissions - Upload"
        ]
      }
    },
    "/permissions/preflight/azure": {
      "post": {
        "description": "**Purpose:** Normalizes Azure credential validation into deployment-ready preflight checks.\n\n**When to call:** Before deployment to get actionable role, subscription, region, and secret-expiration failures.\n\n**Security:** Response is redacted and does not persist credentials.",
        "operationId": "preflightAzureCredentialsFromBody",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AzureCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPreflightResponse"
                }
              }
            },
            "description": "Preflight completed (check ready/status fields)"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The request body failed validation."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Validation Error - The request body failed schema validation"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Run Azure deployment preflight from request body",
        "tags": [
          "Permissions - Upload"
        ]
      }
    },
    "/permissions/preflight/gcp": {
      "post": {
        "description": "**Purpose:** Normalizes GCP credential validation into deployment-ready preflight checks.\n\n**When to call:** Before deployment to get actionable project, billing, API, and region failures.\n\n**Security:** Response is redacted and does not persist credentials.",
        "operationId": "preflightGcpCredentialsFromBody",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GCPCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPreflightResponse"
                }
              }
            },
            "description": "Preflight completed (check ready/status fields)"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The request body failed validation."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Validation Error - The request body failed schema validation"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Run GCP deployment preflight from request body",
        "tags": [
          "Permissions - Upload"
        ]
      }
    },
    "/permissions/verify/aws": {
      "get": {
        "description": "**Purpose:** Debug/local-cloud only. Verifies AWS credentials from project's config_credentials.json.\n\n**Gate:** Disabled unless `ENABLE_LOCAL_CREDENTIAL_FILE_CHECKS=true` is set.\n\n**When to call:** To re-verify deployed project credentials.\n\n**Query param:** project (required) - project context to read.",
        "operationId": "verifyAwsCredentialsFromConfig",
        "parameters": [
          {
            "description": "Project name to load credentials from.",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name to load credentials from.",
              "title": "Project"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsCheckResponse"
                }
              }
            },
            "description": "Credential validation completed (check status field)"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The requested operation is not permitted in this runtime."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Forbidden - The requested operation or resource is protected"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The requested project or resource was not found."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Not Found - The requested resource does not exist"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Verify AWS permissions from project config",
        "tags": [
          "Permissions - Project"
        ]
      },
      "post": {
        "description": "**Purpose:** Verifies AWS credentials against all required permissions for the deployer.\n\n**When to call:** During wizard to validate user-provided AWS credentials.\n\n**Response:** Categorized results by layer (1-5) and by service.",
        "operationId": "verifyAwsCredentialsFromBody",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AWSCredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialsCheckResponse"
                }
              }
            },
            "description": "Credential validation completed (check status field)"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "The request body failed validation."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Validation Error - The request body failed schema validation"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error. Check logs."
                },
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error - An unexpected error occurred"
          }
        },
        "summary": "Verify AWS permissions from request body",
        "tags": [
          "Permissions - Upload"
        ]
      }
    },
...
  },
...
}